Command Updates

+++ src/commands/api-mesh/source/discover.js
@@ -9,12 +9,12 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-const { Command, CliUx } = require('@oclif/core');
-const { promptConfirm, initRequestId } = require('../../../helpers');
-const GetCommand = require('./get');
+const { Command, CliUx, Flags } = require('@oclif/core');
+const { promptConfirm, initRequestId, promptMultiselect, promptSelect } = require('../../../helpers');
 const SourceRegistryStorage = require('source-registry-storage-adapter');
 const config = require('@adobe/aio-lib-core-config');
 const logger = require('../../../classes/logger');
+const InstallCommand = require('./install');
 
 
 class DiscoverCommand extends Command {
@@ -23,6 +23,7 @@ class DiscoverCommand extends Command {
 			await initRequestId();
 
 			logger.info(`RequestId: ${global.requestId}`);
+			const { flags } = await this.parse(DiscoverCommand);
 			const srs = new SourceRegistryStorage(config.get('api-mesh.sourceRegistry.path'));
 			let list;
 			try {
@@ -33,10 +34,17 @@ class DiscoverCommand extends Command {
 			}
 
 			this.generateSourcesTable(list);
-			const needInstall = await promptConfirm(`Are you want to install sources?`);
-			if (needInstall) {
-				GetCommand.run(['-m']);
+			if (!flags.confirm) {
+				const needInstall = await promptConfirm(`Are you want to install sources?`);
+				if (needInstall) {
+					const toInstall = await this.handleMultiple(list)
+					InstallCommand.run([toInstall.join(' ')]);
+				}
+			} else {
+				const toInstall = await this.handleMultiple(list)
+				InstallCommand.run([toInstall.join(' ')]);
 			}
+			
 		} catch (error) {
 			logger.error(error);
 			this.error(`
@@ -44,6 +52,26 @@ class DiscoverCommand extends Command {
 				${error}
 			`);
 		}
+	}
+
+	async handleMultiple(data) {
+		const result = [];
+		const selectedList = await promptMultiselect(
+			'Select sources to install',
+			Object.values(data).map(elem => ({ name: elem.name, value: elem })),
+		);
+		for (const selected of selectedList) {
+			if (selected.versions.length > 1) {
+				const version = await promptSelect(
+					`Please choose the version of "${selected.name}" source`,
+					selected.versions.map(v => ({ name: `v${v}`, value: `${selected.name}@${v}` })),
+				);
+				result.push(version);
+			} else {
+				result.push(`${selected.name}@${selected.latest}`);
+			}
+		}
+		return result;
 	}
 
 	async generateSourcesTable(data) {
@@ -64,6 +92,14 @@ class DiscoverCommand extends Command {
 		CliUx.Table.table(Object.values(data), columns);
 	}
 }
+
+DiscoverCommand.flags = {
+	confirm: Flags.boolean({
+		char: 'c',
+		description:'Auto confirm install action prompt. CLI will not check ask user to install source.',
+		default: false,
+	}),
+};
 
 DiscoverCommand.description = 'Return the list of avaliable sources';
